Create Bin Packing subproblem (First-Fit algorithm)
This is the simplest subproblem of 3D truck loading problem, it's concerned with minimizing the number of bins used (trucks used) without taking in consideration inventory cost of items or arrival time. First-Fit algorithm is used, it's computational complexity is of the order O(nlogn) with best/worst-case scenarios depending on the geometric distribution of items-trucks compatibility.